Master

+++ TouchGFX/gui/include/gui/model/ModelListener.hpp
@@ -15,7 +15,7 @@ public:
         model = m;
     }
 
-    virtual void setADC1 (int val) {};
+    virtual void setADC1 (float val) {};
     virtual void setADC2 (float val) {};
     virtual void setAh (float val) {};
     virtual void tickCounterUpdated (int val) {};
